Changed event class comment, added explicit casts to avoid warnings, and changed include directories to reflect the File class changing to IO

+++ src/Core/Input.cpp
@@ -13,7 +13,7 @@ namespace crynn
 	
 	void scroll_callback(GLFWwindow* window, double xoffset, double yoffset)
 	{
-		Input::OnMouseScroll.Invoke(yoffset); //Invoke the OnMouseScroll when GLFW reports a mouse scroll
+		Input::OnMouseScroll.Invoke((float)yoffset); //Invoke the OnMouseScroll when GLFW reports a mouse scroll
 	}
 
 	bool Input::GetKey(KeyCode key)
